Refactor function parameters
Change function parameters to const reference when possible to avoid
copy overhead

+++ src/random.cpp
@@ -7,6 +7,6 @@ Random::Random(Seed seed, int min, int max) :
     engine{seed},
     dist{min, max} {}
 
-uint8_t Random::generateNumber() {
+int Random::generateNumber() {
     return dist(engine);
 }
